requirements.txt
streamlit
torch
torchvision
timm
opencv-python-headless
efficientnet_pytorch
numpy
Pillow
